src/figures: add includes, .x .y -> float
add .texture, .x, .y, .isWhite
connect Figure to Cell
